+++ CHANGELOG.md
@@ -25,12 +25,15 @@
     * `chaosk8s.crd.actions.apply_from_yaml`
     * `chaosk8s.chaosmesh.stress.actions.stress_cpu`
     * `chaosk8s.chaosmesh.stress.actions.stress_memory`
+    * `chaosk8s.chaosmesh.stress.actions.delete_stressor`
     * `chaosk8s.chaosmesh.network.actions.add_latency`
     * `chaosk8s.chaosmesh.network.actions.corrupt_packets`
     * `chaosk8s.chaosmesh.network.actions.duplicate_packets`
     * `chaosk8s.chaosmesh.network.actions.reorder_packets`
     * `chaosk8s.chaosmesh.network.actions.set_bandwidth`
     * `chaosk8s.chaosmesh.network.actions.set_loss`
+    * `chaosk8s.chaosmesh.network.actions.set_bandwidth`
+    * `chaosk8s.chaosmesh.network.actions.delete_network_fault`
 * Probes:
     * `chaosk8s.secret.probes.secret_exists`
     * `chaosk8s.namespace.probes.namespace_exists`